Refactor v2 workspace page
