+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useClearActivePaneAttention/index.ts
@@ -1,0 +1,1 @@
+export { useClearActivePaneAttention } from "./useClearActivePaneAttention";
