+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useDefaultPaneActions/index.ts
@@ -1,0 +1,1 @@
+export { useDefaultPaneActions } from "./useDefaultPaneActions";
