+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useDirtyTabCloseGuard/index.ts
@@ -1,0 +1,1 @@
+export { useDirtyTabCloseGuard } from "./useDirtyTabCloseGuard";
