+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useWorkspaceFileNavigation/index.ts
@@ -1,0 +1,1 @@
+export { useWorkspaceFileNavigation } from "./useWorkspaceFileNavigation";
