+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useWorkspacePaneOpeners/index.ts
@@ -1,0 +1,1 @@
+export { useWorkspacePaneOpeners } from "./useWorkspacePaneOpeners";
